refactor: simplify testing module structure
