+++ source/geh_calculated_measurements/tests/subsystem_tests/capacity_settlement/seed_table.py
@@ -2,7 +2,7 @@ from datetime import datetime, timezone
 
 from geh_common.domain.types import MeteringPointType, OrchestrationType
 
-from geh_calculated_measurements.testing.utilities.job_tester import JobTestFixture
+from geh_calculated_measurements.testing import JobTestFixture
 from tests import CalculationType, create_random_metering_point_id
 from tests.subsystem_tests import seed_gold_table
 from tests.subsystem_tests.seed_gold_table import GoldTableRow
