+++ source/geh_calculated_measurements/tests/subsystem_tests/capacity_settlement/test_jobs.py
@@ -2,7 +2,7 @@ import uuid
 
 import pytest
 
-from geh_calculated_measurements.testing.utilities.job_tester import JobTest, JobTestFixture
+from geh_calculated_measurements.testing import JobTest, JobTestFixture
 from tests.subsystem_tests.capacity_settlement.seed_table import seed_table
 from tests.subsystem_tests.environment_configuration import EnvironmentConfiguration
