+++ source/geh_calculated_measurements/tests/subsystem_tests/missing_measurements_log/seed_table.py
@@ -3,7 +3,7 @@ from datetime import datetime, timedelta, timezone
 
 from geh_common.domain.types import MeteringPointResolution, MeteringPointType, OrchestrationType
 
-from geh_calculated_measurements.testing.utilities.job_tester import JobTestFixture
+from geh_calculated_measurements.testing import JobTestFixture
 from tests import CalculationType, create_random_metering_point_id
 from tests.subsystem_tests import seed_gold_table
 from tests.subsystem_tests.seed_gold_table import GoldTableRow
